requirements.txt
datetime
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
discord.py
